+++ bundler/text.py
