+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "bulk"
+name = "bundler"
 version = "0.1.0"
 description = ""
 authors = ["JulesBelveze <jules.belveze@hotmail.fr>"]
